Dashboard: make site-launch-modal/index export the generic SiteLaunchModal
Flip the folder so its default export (index.tsx) is the variant-aware
SiteLaunchModal, and move the celebration behavior wrapper into
celebration-modal.tsx (SiteLaunchCelebrationModal). Consumers of the
celebration wrapper now import .../site-launch-modal/celebration-modal,
while the launch hook imports the generic from the folder root. Test
filenames follow their component files.
